+++ loaforcsSoundAPI/LethalCompany/Conditions/DungeonConditionProvider.cs
@@ -10,7 +10,7 @@ internal class DungeonConditionProvider : ConditionProvider
 {
     public override bool Evaluate(SoundReplaceGroup group, JObject conditionDef)
     {
-        SoundPlugin.logger.LogDebug("LethalCompany:dungeon_name value: " + RoundManager.Instance.dungeonGenerator.Generator.DungeonFlow.name);
+        SoundPlugin.logger.LogExtended("LethalCompany:dungeon_name value: " + RoundManager.Instance.dungeonGenerator.Generator.DungeonFlow.name);
         return conditionDef["value"].Value<string>() == RoundManager.Instance.dungeonGenerator.Generator.DungeonFlow.name;
     }
 }