+++ loaforcsSoundAPI/Patches/GameObjectPatch.cs
@@ -18,5 +18,6 @@ internal class GameObjectPatch {
 
         AudioSourceReplaceHelper ext = __instance.AddComponent<AudioSourceReplaceHelper>();
         ext.source = source;
+        SoundPlugin.logger.LogLosingIt("Handled AudioSource created via .AddComponent()");
     }
 }