+++ loaforcsSoundAPI/Providers/Conditions/NotCondition.cs
@@ -8,6 +8,6 @@ using System.Text;
 namespace loaforcsSoundAPI.Providers.Conditions;
 internal class NotCondition : ConditionProvider {
     public override bool Evaluate(SoundReplaceGroup pack, JObject conditionDef) {
-        return !SoundReplacementAPI.ConditionProviders[(string)conditionDef["condition"]["type"]].Evaluate(pack, conditionDef["condition"] as JObject);
+        return !SoundAPI.ConditionProviders[(string)conditionDef["condition"]["type"]].Evaluate(pack, conditionDef["condition"] as JObject);
     }
 }