+++ loaforcsSoundAPI/SoundPlugin.cs
@@ -1,4 +1,5 @@
-﻿using BepInEx;
+﻿using System;
+using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
 using loaforcsSoundAPI.API;
@@ -7,29 +8,31 @@ using loaforcsSoundAPI.Data;
 using loaforcsSoundAPI.LethalCompany;
 using loaforcsSoundAPI.Providers.Conditions;
 using loaforcsSoundAPI.Providers.Formats;
-using loaforcsSoundAPI.Providers.Random;
 using loaforcsSoundAPI.Utils;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using System.Threading;
+using System.Reflection;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+using Random = Unity.Mathematics.Random;
 
 namespace loaforcsSoundAPI {
     [BepInPlugin(MyPluginInfo.PLUGIN_GUID, MyPluginInfo.PLUGIN_NAME, MyPluginInfo.PLUGIN_VERSION)]
     public class SoundPlugin : BaseUnityPlugin {
         public static SoundPlugin Instance { get; private set; }
-        internal static ManualLogSource logger { get; private set; }
-
-        internal static List<string> UniqueSounds;
+        internal static LoafLogger logger { get; private set; }
 
         internal static List<SoundPack> SoundPacks = new List<SoundPack>();
-
+        
+        [Obsolete]
+        internal new static ManualLogSource Logger => logger;
+        
         internal JoinableThreadPool nonstartupThreadPool;
-
+        
         private void Awake() {
-            logger = BepInEx.Logging.Logger.CreateLogSource(MyPluginInfo.PLUGIN_GUID);
+            logger = new LoafLogger(MyPluginInfo.PLUGIN_GUID);
+            BepInEx.Logging.Logger.Sources.Add(logger);
             Instance = this;
 
             logger.LogInfo("Patching...");
@@ -37,26 +40,23 @@ namespace loaforcsSoundAPI {
 
             logger.LogInfo("Setting up config...");
             new SoundPluginConfig(Config);
+            logger.LogInfo("Logging Level at: " + SoundPluginConfig.LOGGING_LEVEL.Value);
 
             logger.LogInfo("Searching for soundpacks...");
             string[] subdirectories = Directory.GetDirectories(Paths.PluginPath);
 
             logger.LogInfo("Beginning Bindings");
             logger.LogInfo("Bindings => General => Audio Formats");
-            SoundReplacementAPI.RegisterAudioFormatProvider(".mp3", new Mp3AudioFormat());
-            SoundReplacementAPI.RegisterAudioFormatProvider(".ogg", new OggAudioFormat());
-            SoundReplacementAPI.RegisterAudioFormatProvider(".wav", new WavAudioFormat());
-
-            logger.LogInfo("Bindings => General => Random Generators");
-            SoundReplacementAPI.RegisterRandomProvider("pure", new PureRandomProvider());
-            SoundReplacementAPI.RegisterRandomProvider("deterministic", new DeterminsticRandomProvider());
+            SoundAPI.RegisterAudioFormatProvider(".mp3", new Mp3AudioFormat());
+            SoundAPI.RegisterAudioFormatProvider(".ogg", new OggAudioFormat());
+            SoundAPI.RegisterAudioFormatProvider(".wav", new WavAudioFormat());
 
             logger.LogInfo("Bindings => General => Conditions");
-            SoundReplacementAPI.RegisterConditionProvider("config", new ConfigCondition());
-            SoundReplacementAPI.RegisterConditionProvider("mod_installed", new ModInstalledConditionProvider());
-            SoundReplacementAPI.RegisterConditionProvider("and", new AndCondition());
-            SoundReplacementAPI.RegisterConditionProvider("not", new NotCondition());
-            SoundReplacementAPI.RegisterConditionProvider("or", new OrCondition());
+            SoundAPI.RegisterConditionProvider("config", new ConfigCondition());
+            SoundAPI.RegisterConditionProvider("mod_installed", new ModInstalledConditionProvider());
+            SoundAPI.RegisterConditionProvider("and", new AndCondition());
+            SoundAPI.RegisterConditionProvider("not", new NotCondition());
+            SoundAPI.RegisterConditionProvider("or", new OrCondition());
 
 
             LethalCompanyBindings.Bind();
@@ -65,7 +65,7 @@ namespace loaforcsSoundAPI {
                 // Combine the current subdirectory with the file name
                 foreach (string subsubdirectory in Directory.GetDirectories(subdirectory)) {
                     string subfilePath = Path.Combine(subsubdirectory, "sound_pack.json");
-                    Logger.LogDebug(subsubdirectory);
+                    logger.LogDebug(subsubdirectory);
 
                     // Check if the file exists
                     if (File.Exists(subfilePath)) {
@@ -83,26 +83,30 @@ namespace loaforcsSoundAPI {
             }
 
             logger.LogInfo("Starting up JoinableThreadPool.");
-            nonstartupThreadPool = new JoinableThreadPool(SoundPluginConfig.THREADPOOL_MAX_THREADS.Value);
 
+            if (SoundPluginConfig.ENABLE_MULTITHREADING.Value)
+                nonstartupThreadPool = new JoinableThreadPool(SoundPluginConfig.THREADPOOL_MAX_THREADS.Value);
+            
             foreach(SoundPack pack in SoundPacks) {
-                pack.QueueNonStartupOnThreadPool(nonstartupThreadPool);
+                pack.QueueLoad(nonstartupThreadPool);
             }
-
-            nonstartupThreadPool.Start();
-            if(!SoundPluginConfig.ENABLE_MULTITHREADING.Value) {
-                logger.LogInfo("Multithreading is disabled :(, joining the thread pool and blocking the main thread.");
-                nonstartupThreadPool.Join();
-            }
+            
+            if(nonstartupThreadPool != null) nonstartupThreadPool.Start();
 
             logger.LogInfo("Registering onSceneLoaded");
 
             SceneManager.sceneLoaded += (Scene scene, LoadSceneMode __) => {
+                logger.LogLosingIt("NEW SCENE LOADED! Scanning for inital playOnAwake sounds...");
+                logger.LogLosingIt($"scene.name: {scene.name}");
+                
                 foreach(AudioSource source in FindObjectsOfType<AudioSource>(true)) {
                     if(source.gameObject.scene != scene) continue; // already processed
 
-
                     if(source.playOnAwake) {
+                        logger.LogLosingIt($"source: {source}");
+                        logger.LogLosingIt($"source.gameObject: {source.gameObject}");
+                        logger.LogLosingIt($"source.clip: {source.clip == null}");
+                        logger.LogLosingIt($"{source.gameObject.name}:"+(source.clip == null? "null" : source.clip.name)+" calling Stop() because its play on awake.");
                         source.Stop();
                     }
 
@@ -110,12 +114,16 @@ namespace loaforcsSoundAPI {
                     ext.source = source;
                 }
             };
-
+            
             logger.LogInfo($"{MyPluginInfo.PLUGIN_GUID}:{MyPluginInfo.PLUGIN_VERSION} has loaded!");
         }
-
-        void OnDisable() {
+        
+        internal void EnsureSoundsAreLoaded() {
+            if(!SoundPluginConfig.ENABLE_MULTITHREADING.Value) return; // if multithreading is disabled, sounds have already been loaeded.
+            
+            logger.LogInfo("Ensuring all sounds are loaded...");
             nonstartupThreadPool.Join();
+            logger.LogInfo("it's all good :3");
         }
     }
 }